requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
gunicorn==21.2.0
flask-sqlalchemy==3.0.5
flask-login==0.6.3
flask-mail==0.9.1
requests==2.31.0
python-dotenv==1.0.0
redis==4.6.0
celery==5.3.4
stripe==7.8.0
psutil==5.9.6
openai==1.3.0
anthropic==0.7.0
httpx==0.25.0
aiohttp==3.9.0
pydub==0.25.1
librosa==0.10.1
soundfile==0.12.1
pyttsx3==2.98
nltk==3.8.1
textstat==0.7.3
langdetect==1.0.9
pandas==2.1.3
numpy==1.24.3
python-magic==0.4.27
pytest==7.4.3
pytest-flask==1.3.0
pytest-mock==3.12.0
factory-boy==3.3.0
flake8==6.1.0
black==23.10.1
isort==5.12.0
